move most code of the scripts to src
